+++ CHANGELOG.md
@@ -301,3 +301,4 @@
 - 2024-02-21: Log follow up, daemon, conf, mutate, +4
 - 2024-02-21: Bench follow up, tz country, sender, configuration, +5
 - 2024-02-22: Add lib, task, follow up, configuration, +2 db
+- 2024-02-23: Metrics discovery, mutate, follow up, qualify, +2